backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
